feat: add event update application service
register, update는 manager가 필요한 "다른 애그리거트"가 요구되는 작업이다.
그러므로 애플리케이션 서비스에 해당한다.

이처럼 다른 애그리거트의 참조가 필요한 서비스는 애플리케이션 서비스로 둔다

+++ event/src/main/java/org/codenbug/event/domain/SeatPriceRange.java
@@ -8,13 +8,13 @@ import lombok.Getter;
 @Getter
 public class SeatPriceRange {
 	@Column(name = "minPrice", nullable = false)
-	private double minPrice;
+	private Integer minPrice;
 	@Column(name = "maxPrice", nullable = false)
-	private double maxPrice;
+	private Integer maxPrice;
 
 	protected SeatPriceRange() {}
 
-	public SeatPriceRange(double minPrice, double maxPrice) {
+	public SeatPriceRange(Integer minPrice, Integer maxPrice) {
 		if( minPrice > maxPrice) {
 			throw new IllegalArgumentException("minPrice must be less than maxPrice");
 		}
